Use pure targets for dynamic imports
Convert dynamic imports in pure modules to pure implementations where available, preserve glTF animation registration through an explicit opt-in, and enforce the pattern in ESLint.
