+++ packages/dev/loaders/src/glTF/2.0/glTFLoader.pure.ts
@@ -72,7 +72,7 @@ import { RandomGUID } from "core/Misc/guid";
 import { type Light } from "core/Lights/light";
 import { BoundingInfo } from "core/Culling/boundingInfo";
 import { type AssetContainer } from "core/assetContainer";
-import { type AnimationPropertyInfo } from "./glTFLoaderAnimation";
+import { type AnimationPropertyInfo } from "./glTFLoaderAnimation.pure";
 import { type IObjectInfo } from "core/ObjectModel/objectModelInterfaces";
 import { registeredGLTFExtensions, registerGLTFExtension, unregisterGLTFExtension, type GLTFExtensionFactory } from "./glTFLoaderExtensionRegistry";
 import { type IInterpolationPropertyInfo } from "core/FlowGraph/typeDefinitions";
@@ -83,8 +83,8 @@ import { Lazy } from "core/Misc/lazy";
 import { type IMaterialLoadingAdapter } from "./materialLoadingAdapter";
 
 // Caching these dynamic imports gives a surprising perf boost (compared to importing them directly each time).
-const LazyAnimationGroupModulePromise = /*#__PURE__*/ new Lazy(() => import("core/Animations/animationGroup"));
-const LazyLoaderAnimationModulePromise = /*#__PURE__*/ new Lazy(() => import("./glTFLoaderAnimation"));
+const LazyAnimationGroupModulePromise = /*#__PURE__*/ new Lazy(() => import("core/Animations/animationGroup.pure"));
+const LazyLoaderAnimationModulePromise = /*#__PURE__*/ new Lazy(() => import("./glTFLoaderAnimation.pure"));
 
 export { GLTFFileLoader };
 
@@ -1864,7 +1864,8 @@ export class GLTFLoader implements IGLTFLoader {
         }
 
         // async-load the animation sampler to provide the interpolation of the channelTargetPath
-        return LazyLoaderAnimationModulePromise.value.then(() => {
+        return LazyLoaderAnimationModulePromise.value.then(({ RegisterGLTFLoaderAnimation: registerGLTFLoaderAnimation }) => {
+            registerGLTFLoaderAnimation();
             let properties: IInterpolationPropertyInfo[];
             switch (channelTargetPath) {
                 case AnimationChannelTargetPath.TRANSLATION: {
