+++ packages/dev/loaders/src/glTF/2.0/pure.ts
@@ -4,6 +4,7 @@ export * from "./glTFLoader.pure";
 export * from "./glTFLoaderExtension";
 export * from "./glTFLoaderExtensionRegistry";
 export * from "./glTFLoaderInterfaces";
+export * from "./glTFLoaderAnimation.pure";
 export * from "./openpbrMaterialLoadingAdapter";
 export * from "./pbrMaterialLoadingAdapter";
 export * from "./Extensions/pure";
